Move stream_compression to src/core/lib/compression
