+++ config.w32
@@ -71,6 +71,7 @@ if (PHP_GRPC != "no") {
     "src\\core\\lib\\channel\\handshaker_registry.c " +
     "src\\core\\lib\\compression\\compression.c " +
     "src\\core\\lib\\compression\\message_compress.c " +
+    "src\\core\\lib\\compression\\stream_compression.c " +
     "src\\core\\lib\\http\\format_request.c " +
     "src\\core\\lib\\http\\httpcli.c " +
     "src\\core\\lib\\http\\parser.c " +
@@ -205,7 +206,6 @@ if (PHP_GRPC != "no") {
     "src\\core\\ext\\transport\\chttp2\\transport\\huffsyms.c " +
     "src\\core\\ext\\transport\\chttp2\\transport\\incoming_metadata.c " +
     "src\\core\\ext\\transport\\chttp2\\transport\\parsing.c " +
-    "src\\core\\ext\\transport\\chttp2\\transport\\stream_compression.c " +
     "src\\core\\ext\\transport\\chttp2\\transport\\stream_lists.c " +
     "src\\core\\ext\\transport\\chttp2\\transport\\stream_map.c " +
     "src\\core\\ext\\transport\\chttp2\\transport\\varint.c " +
